refactor(multi_object_tracker): restructure overlap merge into a staged pipeline
Extract the tracker overlap merge into a merger/ module: shared snapshot and
decision-context types (detail/overlap_types), the spatial gate (detail/
overlap_gate, per-label pruning distances), and the survival ranking
(detail/survival_ranking). tracker_overlap_manager drives a four-stage
pipeline — snapshot, decide directed merge edges, group one best winner per
loser, apply the star-forest merges. Move redundancy_check under merger/detail
and update includes.
