+++ perception/autoware_multi_object_tracker/CMakeLists.txt
@@ -42,11 +42,13 @@ set(${PROJECT_NAME}_lib
   lib/odometry.cpp
   lib/association/bev_association.cpp
   lib/association/association_manager.cpp
-  lib/association/tracker_overlap_manager.cpp
+  lib/merger/tracker_overlap_manager.cpp
+  lib/merger/survival_ranking.cpp
+  lib/merger/overlap_gate.cpp
   lib/association/polar_association.cpp
   lib/association/scoring/bev_assignment_scoring.cpp
   lib/association/scoring/polar_assignment_scoring.cpp
-  lib/association/scoring/redundancy_check.cpp
+  lib/merger/redundancy_check.cpp
   lib/association/mu_successive_shortest_path/mu_ssp.cpp
   lib/types.cpp
   lib/object_model/uuid.cpp
