+++ perception/autoware_multi_object_tracker/src/processor/processor.hpp
@@ -17,7 +17,7 @@
 
 #include "autoware/multi_object_tracker/association/adaptive_threshold_cache.hpp"
 #include "autoware/multi_object_tracker/association/association_manager.hpp"
-#include "autoware/multi_object_tracker/association/tracker_overlap_manager.hpp"
+#include "autoware/multi_object_tracker/merger/tracker_overlap_manager.hpp"
 #include "autoware/multi_object_tracker/tracker/trackers/tracker_base.hpp"
 #include "autoware/multi_object_tracker/types.hpp"
